Address concerns
