+++ lib/reply.js
@@ -194,7 +194,7 @@ function handleError (reply, error, cb) {
   if (statusCode >= 500) {
     reply.res.log.error({ req: reply.request.raw, res: reply.res, err: error }, error && error.message)
   } else if (statusCode >= 400) {
-    reply.res.log.info({ req: reply.request.raw, res: reply.res, err: error }, error && error.message)
+    reply.res.log.info({ res: reply.res, err: error }, error && error.message)
   }
 
   if (error && error.headers) {
