+++ package.json
@@ -73,7 +73,6 @@
     "dns-prefetch-control": "^0.1.0",
     "fast-json-body": "^1.1.0",
     "fastify-plugin": "^0.2.1",
-    "flush-write-stream": "^1.0.2",
     "frameguard": "^3.0.0",
     "h2url": "^0.1.2",
     "helmet": "^3.9.0",
